+++ digiwf-integrations/digiwf-cosys-integration/digiwf-cosys-integration-core/src/main/java/de/muenchen/oss/digiwf/cosys/integration/application/port/in/CreateDocumentInPort.java
@@ -1,6 +1,6 @@
 package de.muenchen.oss.digiwf.cosys.integration.application.port.in;
 
-import de.muenchen.oss.digiwf.cosys.integration.model.GenerateDocument;
+import de.muenchen.oss.digiwf.cosys.integration.domain.model.GenerateDocument;
 import jakarta.validation.Valid;
 
 public interface CreateDocumentInPort {
