+++ digiwf-integrations/digiwf-cosys-integration/digiwf-cosys-integration-core/src/main/java/de/muenchen/oss/digiwf/cosys/integration/domain/model/DocumentStorageUrl.java
@@ -1,19 +1,17 @@
-package de.muenchen.oss.digiwf.email.integration.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
+package de.muenchen.oss.digiwf.cosys.integration.domain.model;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
 
-/**
- * Attachment File you want to get from the S3 storage.
- */
 @Data
+@Builder
 @RequiredArgsConstructor
 @AllArgsConstructor
-public class PresignedUrl {
+public class DocumentStorageUrl {
 
     /**
      * Url to the s3 service.
@@ -29,10 +27,10 @@ public class PresignedUrl {
 
     /**
      * Proper Http Method (Post, Put, Get, Delete) to interact with S3.
-     * Note: Only GET is supported. The mail integration is not intended to modify files!
+     * Note: Only POST and PUT is supported.
      */
     @NotBlank(message = "Action is mandatory")
-    @Pattern(regexp = "GET", message = "Only action GET is supported")
+    @Pattern(regexp = "^(POST|PUT)$", message = "Only action POST or PUT is supported")
     private String action;
 
 }
