+++ digiwf-integrations/digiwf-cosys-integration/digiwf-cosys-integration-core/src/main/java/de/muenchen/oss/digiwf/cosys/integration/domain/model/GenerateDocumentRequest.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.digiwf.cosys.integration.model;
+package de.muenchen.oss.digiwf.cosys.integration.domain.model;
 
 import lombok.Data;
