+++ digiwf-integrations/digiwf-cosys-integration/digiwf-cosys-integration-core/src/test/java/de/muenchen/oss/digiwf/cosys/integration/domain/model/GenerateDocumentTest.java
@@ -3,8 +3,6 @@ package de.muenchen.oss.digiwf.cosys.integration.domain.model;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.muenchen.oss.digiwf.cosys.integration.model.DocumentStorageUrl;
-import de.muenchen.oss.digiwf.cosys.integration.model.GenerateDocument;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
