+++ digiwf-integrations/digiwf-cosys-integration/digiwf-cosys-integration-starter/src/main/java/de/muenchen/oss/digiwf/cosys/integration/configuration/CosysAutoConfiguration.java
@@ -11,7 +11,7 @@ import de.muenchen.oss.digiwf.cosys.integration.application.port.in.CreateDocume
 import de.muenchen.oss.digiwf.cosys.integration.application.port.out.GenerateDocumentOutPort;
 import de.muenchen.oss.digiwf.cosys.integration.application.port.out.SaveFileToStorageOutPort;
 import de.muenchen.oss.digiwf.cosys.integration.application.usecase.CreateDocumentUseCase;
-import de.muenchen.oss.digiwf.cosys.integration.model.GenerateDocument;
+import de.muenchen.oss.digiwf.cosys.integration.domain.model.GenerateDocument;
 import de.muenchen.oss.digiwf.message.process.api.ErrorApi;
 import de.muenchen.oss.digiwf.message.process.api.ProcessApi;
 import de.muenchen.oss.digiwf.s3.integration.client.configuration.S3IntegrationClientAutoConfiguration;
