+++ digiwf-integrations/digiwf-email-integration/digiwf-email-integration-core/src/main/java/de/muenchen/oss/digiwf/email/integration/application/port/in/SendMailInPort.java
@@ -1,7 +1,7 @@
 package de.muenchen.oss.digiwf.email.integration.application.port.in;
 
-import de.muenchen.oss.digiwf.email.integration.model.TemplateMail;
-import de.muenchen.oss.digiwf.email.integration.model.TextMail;
+import de.muenchen.oss.digiwf.email.integration.domain.model.TemplateMail;
+import de.muenchen.oss.digiwf.email.integration.domain.model.TextMail;
 import jakarta.validation.Valid;
 
 public interface SendMailInPort {
