+++ digiwf-integrations/digiwf-email-integration/digiwf-email-integration-core/src/main/java/de/muenchen/oss/digiwf/email/integration/application/usecase/SendMailUseCase.java
@@ -3,9 +3,9 @@ package de.muenchen.oss.digiwf.email.integration.application.usecase;
 import de.muenchen.oss.digiwf.email.integration.application.port.in.SendMailInPort;
 import de.muenchen.oss.digiwf.email.integration.application.port.out.LoadMailAttachmentOutPort;
 import de.muenchen.oss.digiwf.email.integration.application.port.out.MailOutPort;
-import de.muenchen.oss.digiwf.email.integration.model.BasicMail;
-import de.muenchen.oss.digiwf.email.integration.model.TemplateMail;
-import de.muenchen.oss.digiwf.email.integration.model.TextMail;
+import de.muenchen.oss.digiwf.email.integration.domain.model.BasicMail;
+import de.muenchen.oss.digiwf.email.integration.domain.model.TemplateMail;
+import de.muenchen.oss.digiwf.email.integration.domain.model.TextMail;
 import de.muenchen.oss.digiwf.email.model.FileAttachment;
 import de.muenchen.oss.digiwf.message.process.api.error.BpmnError;
 import freemarker.template.TemplateException;
