+++ digiwf-integrations/digiwf-email-integration/digiwf-email-integration-example/src/main/java/de/muenchen/oss/digiwf/email/integration/domain/model/PresignedUrl.java
@@ -1,11 +1,10 @@
-package de.muenchen.oss.digiwf.email.integration.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
+package de.muenchen.oss.digiwf.email.integration.domain.model;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
 
 /**
  * Attachment File you want to get from the S3 storage.
